requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]>=3.1
python-dotenv
pydantic
stripe==10.11.0